rad_net / rad_amb / rad_flux go int32 -> int64 together with all 28 surviving
32-bit surfaces in 20 files -- four bindings (not three), both CUDA atomics via
the unsigned-long-long idiom, a CAS-loop sensor twin, two narrowing casts
deleted, and cuda_temperature's own D2H byte count (its nb is shared with eight
still-int32 planes, so reusing it would have copied half a plane silently).
The old cast still writes them: same integers, wider storage.

Behaviour-neutral and verified so: no golden, spec or digest file touched;
suite 2557 passed / 0 failed with the CUDA gates (my own run), CUDA lockstep
tol-0 on all three planes.

THREE FINDINGS, all verified by the orchestrator against the code:

1. rad_flux's INT32_MAX was never an overflow guard -- it is a CAP ON UNIT HEAT
   DAMAGE (its one consumer is exchange.py:330) and it BINDS IN ORDINARY PLAY
   (measured: engages from tick 12, 38 of 120 ticks, up to 249 cells at once,
   un-capped peak 459% of it; my own run of the tripwire sees 157 capped
   cell-ticks in 24 ticks). Saturating at INT64_MAX instead would have made
   fires meaningfully more lethal with no golden able to catch it (rad_flux is
   in neither DIGEST_FIELDS nor SIM_FIELDS). The ceiling stays, now explicit as
   RAD_FLUX_CEILING with a non-vacuous tripwire. WHETHER IT SHOULD EXIST IS AN
   OPEN FEEL QUESTION FOR ERIK, due before P3b.

2. .noconvert() is INERT on a nullable argument: a py::object (needed for the
   None idiom) has no overload pass to annotate, so .cast<>() runs forcecast
   directly. Two of the four bindings would have looked hardened and behaved
   exactly as before; they use explicit isinstance + type_error instead. This
   generalises to every nullable argument in the codebase.

3. The hardening caught a stale caller grep could not:
   cuda_pr1_fire_plane_check.py allocates positionally, so no line names
   rad_net. Without noconvert it would have compared two truncated temporaries
   and PASSED -- a CPU/CUDA lockstep gate silently measuring nothing.

+++ cpp/src/cuda_raycaster.h
@@ -62,8 +62,14 @@ struct RayHD {
 // nullptr (the default) == exchange OFF, i.e. every legacy caller
 // (cuda_raycaster_cast / _cast_batch) marches exactly as before. When supplied,
 // each absorbing marched cell runs the antisymmetric pair update into
-// `rad_net` — plain SIGNED int32 atomicAdd, order-free and exact, both ends
-// getting the SAME integer. `e_table` is the E_TABLE_SIZE-entry black-body bake
+// `rad_net` — order-free and exact, both ends getting the SAME integer. Since
+// P3a-1 the plane is int64 and CUDA has no signed 64-bit atomicAdd, so it is
+// the tree's standard idiom `atomicAdd((unsigned long long*)p, (unsigned long
+// long)v)` (design v3 §8.2; cuda_bulk_transport.cu:313, cuda_combustion.cu,
+// cuda_eos_resident.cu:940): unsigned wrap modulo 2^64 IS two's-complement
+// signed 64-bit addition, so the reinterpreted int64 is exactly the CPU's
+// `rad_signed_add`, and modular addition is associative + commutative, so the
+// accumulation is order-free at 64 bits exactly as it was at 32. `e_table` is the E_TABLE_SIZE-entry black-body bake
 // (uploaded with the rest of the per-call input set — see the .cu).
 //
 // P-F1a: the radiation arguments are GONE from this entry point. The exchange
@@ -96,8 +102,8 @@ void raycaster_cast_directional(
 //
 // `emit_mask` is the once-per-tick emitter plane (v7.1 item 13) — rule 2 keys
 // on it, so both backends read the SAME bytes. `rad_amb` is the per-tile sky
-// ledger: PLAIN int32 atomicAdd, deliberately not one global counter, so the
-// accumulation is order-free and contention-free; the host reduces it.
+// ledger: the same PLAIN int64 atomic, deliberately not one global counter, so
+// the accumulation is order-free and contention-free; the host reduces it.
 //
 // RETURNS: the number of rays that terminated on a CONTACT FACE (rule 3) —
 // the same diagnostic Raycaster::cast_from_fire_plane returns, so the
@@ -112,13 +118,13 @@ int64_t raycaster_cast_radiation(
     const int32_t* temperature,              // Q16.16 (h,w)
     const int32_t* heat_inv_shift,           // (h,w)
     const uint8_t* emit_mask,                // (h,w) 1 == emitter
-    int32_t* rad_net,                        // Q16.16 (h,w) signed accumulator
-    int32_t* rad_amb,                        // (h,w) sky ledger (rule 4)
+    int64_t* rad_net,                        // Q16.16 (h,w) signed accumulator
+    int64_t* rad_amb,                        // (h,w) sky ledger (rule 4)
     // D3: the RADIANT-FLUX SENSOR plane — positive-only, AIR cells only, NOT
     // part of the energy ledger (no temperature effect, nothing debited); its
     // only consumer is unit heat damage. Saturating atomic == the old heat
     // contract, so it is order-free. See raycaster.h RadCtx::rad_flux.
-    int32_t* rad_flux = nullptr);
+    int64_t* rad_flux = nullptr);
 
 // Backend flag (mirrors the S1 temperature backend switch).
 bool raycaster_backend_is_cuda();
